Improved coverage
to dev deps

separated withdrawal tests

remove container after build, update bin

removed unused PartialEq

test

profile moved to cargo config

test products with fee conversion

added lint step, some warnings fixes

install clippy

machete and typos

install toolchain

typos

enabled more lints

some more coverage

folder for common module

moved tests from contract mod.rs
