+++ .cargo/config
@@ -1,2 +1,9 @@
-[build]
-rustflags = ["-C", "link-args=-s"]
+[profile.contract]
+inherits = "release"
+strip = true
+codegen-units = 1
+opt-level = "z"
+lto = true
+debug = false
+panic = "abort"
+overflow-checks = true
