+++ contract/src/claim/mod.rs
@@ -6,7 +6,7 @@ mod tests {
 
     use crate::{
         claim::api::ClaimApi,
-        common::{tests::Context, UDecimal, U32},
+        common::{tests::Context, u32::U32, udecimal::UDecimal},
         jar::{api::JarApi, model::Jar},
         product::{
             model::{Apy, Product},
@@ -39,17 +39,15 @@ mod tests {
         let admin = accounts(1);
 
         let product = generate_product();
-        let reference_jar = Jar::generate(0, &alice, &product.id).principal(100_000_000_000);
-        let mut context = Context::new(admin)
-            .with_products(&[product])
-            .with_jars(&[reference_jar.clone()]);
+        let jar = Jar::generate(0, &alice, &product.id).principal(100_000_000_000);
+        let mut context = Context::new(admin).with_products(&[product]).with_jars(&[jar.clone()]);
 
         context.set_block_timestamp_in_days(365);
 
         context.switch_account(&alice);
-        context.contract.claim_jars(vec![reference_jar.index], Some(U128(100)));
+        context.contract.claim_jars(vec![jar.index], Some(U128(100)));
 
-        let jar = context.contract.get_jar(U32(reference_jar.index));
+        let jar = context.contract.get_jar(U32(jar.index));
         assert_eq!(100, jar.claimed_balance.0);
     }
