+++ scripts/build.sh
@@ -4,6 +4,6 @@ set -eox pipefail
 echo ">> Building contract"
 
 rustup target add wasm32-unknown-unknown
-cargo build -p sweat_jar --target wasm32-unknown-unknown --release
+cargo build -p sweat_jar --target wasm32-unknown-unknown --profile=contract
 
-mv ./target/wasm32-unknown-unknown/release/sweat_jar.wasm res/sweat_jar.wasm
+mv ./target/wasm32-unknown-unknown/contract/sweat_jar.wasm res/sweat_jar.wasm
