--- scripts/coverage.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-cargo llvm-cov --hide-instantiations --open
